diy_struct_service select fail!
