+++ db.go
@@ -93,7 +93,7 @@ func OpenTestConnection() (db *gorm.DB, err error) {
 
 func RunMigrations() {
 	var err error
-	allModels := []interface{}{&Channel{}, &User{}, &Form{}}
+	allModels := []interface{}{&Channel{}, &User{}, &Form{}, &Article{}}
 	rand.Seed(time.Now().UnixNano())
 	rand.Shuffle(len(allModels), func(i, j int) { allModels[i], allModels[j] = allModels[j], allModels[i] })
